+++ packages/agents/router/src/lib/helpers/shared.ts
@@ -4,6 +4,7 @@ import {
   signRouterPathPayload as _signRouterPathPayload,
   getSubgraphHealth as _getSubgraphHealth,
   getSubgraphName as _getSubgraphName,
+  recoverRouterPathPayload as _recoverRouterPathPayload,
 } from "@connext/nxtp-utils";
 import { utils } from "ethers";
 
@@ -29,9 +30,9 @@ export const getDestinationLocalAsset = async (
   // get canonical asset from orgin domain.
   const sendingDomainAsset = await subgraph.getAssetByLocal(_originDomain, _originLocalAsset);
 
-  const canonicalId = sendingDomainAsset!.canonicalId as string;
+  const canonicalId = sendingDomainAsset!.canonicalId ;
 
-  const destinationDomainAsset = await subgraph.getAssetByCanonicalId(Number(_destinationDomain), canonicalId);
+  const destinationDomainAsset = await subgraph.getAssetByCanonicalId(_destinationDomain, canonicalId);
 
   const localAddress = destinationDomainAsset!.local;
   return localAddress;
@@ -42,6 +43,7 @@ export const getTransactionId = (nonce: string, domain: string): string => {
 };
 
 export const signRouterPathPayload = _signRouterPathPayload;
+export const recoverRouterPathPayload = _recoverRouterPathPayload;
 
 export const getSubgraphHealth = _getSubgraphHealth;