+++ packages/deployments/contracts/hardhat.config.ts
@@ -115,8 +115,7 @@ const config: HardhatUserConfig = {
       accounts: { mnemonic },
       chainId: 4,
       url: urlOverride || process.env.RINKEBY_ETH_PROVIDER_URL || "http://localhost:8545",
-      minGasPrice: utils.parseUnits("10", "gwei").toString(),
-      gasPrice: utils.parseUnits("10", "gwei").toNumber(),
+      gasPrice: utils.parseUnits("20", "gwei").toNumber(),
     },
     goerli: {
       accounts: { mnemonic },