+++ packages/deployments/contracts/test/connext.spec.ts
@@ -15,6 +15,7 @@ import {
   ConnextHandler,
   ConnextLogic,
   RelayerFeeRouter,
+  PromiseRouter,
 } from "../typechain-types";
 
 import {
@@ -97,6 +98,8 @@ describe("Connext", () => {
   let stableSwap: DummySwap;
   let originRelayerFeeRouter: RelayerFeeRouter;
   let destinationRelayerFeeRouter: RelayerFeeRouter;
+  let originPromiseRouter: PromiseRouter;
+  let destinationPromiseRouter: PromiseRouter;
   let home: Home;
   let destinationHome: Home;
   let snapshot: number;
@@ -149,6 +152,15 @@ describe("Connext", () => {
       [destinationXappConnectionManager.address],
     );
 
+    // Deploy PromiseRouters
+    originPromiseRouter = await deployUpgradeableProxy<PromiseRouter>("PromiseRouter", proxyOwner.address, [
+      originXappConnectionManager.address,
+    ]);
+
+    destinationPromiseRouter = await deployUpgradeableProxy<PromiseRouter>("PromiseRouter", proxyOwner.address, [
+      destinationXappConnectionManager.address,
+    ]);
+
     // Deploy bridge
     originBridge = (
       await deployUpgradeableProxy<ConnextHandler>(
@@ -160,6 +172,7 @@ describe("Connext", () => {
           originTokenRegistry.address,
           weth.address,
           originRelayerFeeRouter.address,
+          originPromiseRouter.address,
         ],
         {
           ConnextLogic: ConnextLogic.address,
@@ -178,6 +191,7 @@ describe("Connext", () => {
           destinationTokenRegistry.address,
           weth.address,
           destinationRelayerFeeRouter.address,
+          destinationPromiseRouter.address,
         ],
         {
           ConnextLogic: ConnextLogic.address,
@@ -846,6 +860,10 @@ describe("Connext", () => {
       callData: "0x",
       originDomain,
       destinationDomain,
+      callback: ZERO_ADDRESS,
+      callbackFee: 0,
+      forceSlow: false,
+      receiveLocal: false,
     };
     const transactingAssetId = originAdopted.address;
     const amount = utils.parseEther("0.0001");
@@ -940,6 +958,10 @@ describe("Connext", () => {
       callData: "0x",
       originDomain,
       destinationDomain,
+      callback: ZERO_ADDRESS,
+      callbackFee: 0,
+      forceSlow: false,
+      receiveLocal: false,
     };
     const transactingAssetId = constants.AddressZero;
     const amount = utils.parseEther("0.0001");
@@ -1043,6 +1065,10 @@ describe("Connext", () => {
       callData: "0x",
       originDomain,
       destinationDomain,
+      callback: ZERO_ADDRESS,
+      callbackFee: 0,
+      forceSlow: false,
+      receiveLocal: false,
     };
     const transactingAssetId = originAdopted.address;
     const amount = utils.parseEther("0.0001");
@@ -1120,6 +1146,10 @@ describe("Connext", () => {
       callData: "0x",
       originDomain,
       destinationDomain,
+      callback: ZERO_ADDRESS,
+      callbackFee: 0,
+      forceSlow: false,
+      receiveLocal: false,
     };
     const amount = utils.parseEther("0.001");
     let message: any;
@@ -1338,6 +1368,10 @@ describe("Connext", () => {
       callData: "0x",
       originDomain,
       destinationDomain,
+      callback: ZERO_ADDRESS,
+      callbackFee: 0,
+      forceSlow: false,
+      receiveLocal: false,
     };
 
     beforeEach(async () => {